게시물 작성하기 추가 UI, 유효성 검사 수정 및 게시물 신고하기 수정 완료
