+++ livingalone/lib/main.dart
@@ -10,8 +10,9 @@ import 'package:livingalone/handover/view/add_room_handover_screen6.dart';
 import 'package:livingalone/handover/view/add_room_handover_screen7.dart';
 import 'package:livingalone/handover/view/add_room_handover_screen8.dart';
 import 'package:livingalone/handover/view/add_ticket_handover_screen1.dart';
-import 'package:livingalone/handover/view/add_ticket_handover_screen3.dart';
-import 'package:livingalone/handover/view/add_ticket_handover_screen5.dart';
+import 'package:livingalone/handover/view/add_ticket_handover_screen2.dart';
+import 'package:livingalone/handover/view/add_ticket_handover_screen4.dart';
+import 'package:livingalone/handover/view/add_ticket_handover_screen6.dart';
 import 'package:livingalone/post_modify/view/edit_room_post_screen.dart';
 import 'package:livingalone/home/view/home_screen.dart';
 import 'package:flutter_screenutil/flutter_screenutil.dart';
@@ -60,9 +61,9 @@ class MyApp extends ConsumerWidget {
             ),
             debugShowCheckedModeBanner: false,
             // home: LivingDetailScreen(postType: PostType.ticket, postId: '10',),
-            home: RootTab()
+            home: ReportScreen()
             // home: AddRoomHandoverScreen7(rentType: RentType.shortRent,),
-            // home: AddRoomHandoverScreen6(),
+            // home: AddTicketHandoverScreen2()
           ),
         );
       }
